requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
pytz==2024.1
feedparser==6.0.10
